move questions listener to app - bug (#348)
